Move custom Ginkgo matchers to matchers package.
It's important for matchers to be in a separate packages. Ginko convention is to
use anonymous imports but that's a bad idea for anything else. Therefore Gingko
extensions and other code can't share a package.
