+++ pkg/apis/logging/v1/cluster_log_forwarder_types_test.go
@@ -9,6 +9,7 @@ import (
 
 	. "github.com/openshift/cluster-logging-operator/pkg/apis/logging/v1"
 	. "github.com/openshift/cluster-logging-operator/test"
+	. "github.com/openshift/cluster-logging-operator/test/matchers"
 )
 
 var _ = Describe("ClusterLogForwarder", func() {
