+++ pkg/generators/forwarding/fluentd/generators_test.go
@@ -3,7 +3,7 @@ package fluentd
 import (
 	"text/template"
 
-	. "github.com/openshift/cluster-logging-operator/test"
+	. "github.com/openshift/cluster-logging-operator/test/matchers"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
