+++ test/matchers/diff_test.go
@@ -1,10 +1,9 @@
-package test_test
+package matchers_test
 
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-
-	. "github.com/openshift/cluster-logging-operator/test"
+	. "github.com/openshift/cluster-logging-operator/test/matchers"
 )
 
 var _ = Describe("Matchers", func() {
